增加性能测试脚本
# message auto-generated for no-merge-commit merge:
merge build into master

[DT]增加server解析性能测试脚本

Created-by: insightDeveloper

Match-id-9eb7783188ced92b292a64f88a705bf31a8741ed

+++ server/build/preprocess_third_party.py
@@ -58,12 +58,13 @@ def log(info):
 def prepare_sqlite_src():
     log('start to prepare sqlite src')
     if platform.system() == "Windows":
-        tar_path = os.path.join(THIRD_PARTY_DIR, SQLITE_SRC_TAR)
-        urllib.request.urlretrieve(SQLITE3_SOURCE_URL, tar_path)
-        tar = tarfile.open(tar_path)
-        tar.extractall(THIRD_PARTY_DIR)
-        tar.close()
-        os.remove(tar_path)
+        if not os.path.exists(os.path.join(THIRD_PARTY_DIR, SQLITE3_AUTOCONF_DIR)):
+            tar_path = os.path.join(THIRD_PARTY_DIR, SQLITE_SRC_TAR)
+            urllib.request.urlretrieve(SQLITE3_SOURCE_URL, tar_path)
+            tar = tarfile.open(tar_path)
+            tar.extractall(THIRD_PARTY_DIR)
+            tar.close()
+            os.remove(tar_path)
     else:
         build_path = os.path.join(THIRD_PARTY_DIR, SQLITE3_SRC_DIR, 'build')
         if os.path.exists(build_path):
